issues/29 refactor(Subscriber): rename AddSustainTo AddDestroyTo
Reducer's AddTo is currently purged.
